- Add optional JPEG decoding support for EAF animations - Center QR code rendering in UI layout - Add alpha channel support for animations - Refactor test_app into modular components
