+++ .readthedocs.yaml
@@ -15,4 +15,3 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
-
