+++ test_apps/main/CMakeLists.txt
@@ -2,13 +2,7 @@
 idf_component_register(
     SRC_DIRS "." "./assets"
     INCLUDE_DIRS "."
-)
-
-get_filename_component(ESP_EMOTE_GFX_COMPONENT_DIR "${CMAKE_CURRENT_LIST_DIR}/../.." ABSOLUTE)
-
-# set_property(SOURCE font_puhui_16_4.c APPEND PROPERTY
-#     INCLUDE_DIRECTORIES "${ESP_EMOTE_GFX_COMPONENT_DIR}/include"
-# )
+    WHOLE_ARCHIVE)
 
 spiffs_create_partition_assets(
     test_assets
