+++ test_apps/main/idf_component.yml
@@ -9,7 +9,7 @@ dependencies:
   esp_mmap_assets:
     version: 1.*
 
-  espressif/esp-box-3:
+  espressif/esp-box:
     version: "*"
     rules:
     - if: "${IDF_TARGET} == esp32s3"
